Revert "Public Course Import/Export API (continued)"
