+++ cms/templates/import.html
@@ -53,7 +53,7 @@ else:
 
       </div>
 
-      <form id="fileupload" method="post" action="${import_url}" enctype="multipart/form-data" class="import-form">
+      <form id="fileupload" method="post" enctype="multipart/form-data" class="import-form">
 
         ## Translators: ".tar.gz" is a file extension, and files with that extension are called "gzipped tar files": these terms should not be translated
         <h2 class="title">
