--- openedx/core/djangoapps/import_export/__init__.py
@@ -1,3 +1,0 @@
-"""
-Publishing API
-"""
