--- openedx/core/djangoapps/import_export/courses/__init__.py
@@ -1,3 +1,0 @@
-"""
-Course publishing API
-"""
