--- openedx/core/djangoapps/import_export/courses/tests/__init__.py
@@ -1,3 +1,0 @@
-"""
-Tests for course publishing API
-"""
