--- openedx/core/djangoapps/import_export/models.py
@@ -1,3 +1,0 @@
-"""
-A models.py is required to make this an app (until we move to Django 1.7)
-"""
